2. 뷰 컨트롤러 구성하기 // controller 파일의 많은 책임을 extension을 활용해서 덜어내기
